+++ CHANGES.md
@@ -363,3 +363,5 @@
 - Portable release: rename Win32 to Win7x86, which supports Windows 7 or newer version of Windows OS, both x86 (32-bit) and x64 (64-bit) architecture
 - Portable release: add Win10x64 bundled with [python-3.14.2-embed-amd64](https://www.python.org/ftp/python/3.14.2/python-3.14.2-embed-amd64.zip) package and necessary modules, which supports Windows 10 or newer version of Windows OS, x64 (64-bit) architecture only
 
+### Pre-release (work-in-progress)
+- HIMSA: fix for Noahlink Wireless firmware upgrader filename changes in v3.6.0.0